+++ packages/headless/src/features/product-listing/v2/product-listing-v2-actions.ts
@@ -1,22 +1,19 @@
 import {createAsyncThunk} from '@reduxjs/toolkit';
-import {AsyncThunkProductListingV2Options} from '../../../api/commerce/product-listings/v2/product-listing-v2-api-client';
-import {
-  ProductListingV2Request,
-  ProductListingV2SuccessResponse,
-} from '../../../api/commerce/product-listings/v2/product-listing-v2-request';
+import {AsyncThunkCommerceOptions} from '../../../api/commerce/commerce-api-client';
+import {ProductListingV2Request} from '../../../api/commerce/product-listings/v2/product-listing-v2-request';
+import {ProductListingV2SuccessResponse} from '../../../api/commerce/product-listings/v2/product-listing-v2-response';
 import {isErrorResponse} from '../../../api/search/search-api-client';
 import {
   CategoryFacetSection,
   ConfigurationSection,
-  ContextSection,
   DateFacetSection,
-  FacetOptionsSection,
   FacetOrderSection,
   FacetSection,
   NumericFacetSection,
   PaginationSection,
   ProductListingV2Section,
   StructuredSortSection,
+  VersionSection,
 } from '../../../state/state-sections';
 import {sortFacets} from '../../../utils/facet-utils';
 import {
@@ -30,15 +27,14 @@ import {logProductListingV2Load} from './product-listing-v2-analytics';
 export type StateNeededByFetchProductListingV2 = ConfigurationSection &
   ProductListingV2Section &
   Partial<
-    PaginationSection &
-      StructuredSortSection &
-      FacetSection &
+    FacetSection &
       NumericFacetSection &
       CategoryFacetSection &
       DateFacetSection &
-      FacetOptionsSection &
       FacetOrderSection &
-      ContextSection
+      StructuredSortSection &
+      PaginationSection &
+      VersionSection
   >;
 
 export interface FetchProductListingV2ThunkReturn {
@@ -53,13 +49,13 @@ export interface FetchProductListingV2ThunkReturn {
 export const fetchProductListing = createAsyncThunk<
   FetchProductListingV2ThunkReturn,
   void,
-  AsyncThunkProductListingV2Options<StateNeededByFetchProductListingV2>
+  AsyncThunkCommerceOptions<StateNeededByFetchProductListingV2>
 >(
-  'productlisting/v2/fetch',
+  'commerce/productListing/fetch',
   async (_action, {getState, dispatch, rejectWithValue, extra}) => {
     const state = getState();
     const {apiClient} = extra;
-    const fetched = await apiClient.getListing(
+    const fetched = await apiClient.getProductListing(
       await buildProductListingRequestV2(state)
     );
 
@@ -82,11 +78,11 @@ export const buildProductListingRequestV2 = async (
 
   return {
     accessToken: state.configuration.accessToken,
-    platformUrl: state.configuration.platformUrl,
+    url: state.configuration.platformUrl,
     organizationId: state.configuration.organizationId,
     trackingId: state.productListing.trackingId,
-    locale: state.productListing.locale,
-    mode: state.productListing.mode,
+    language: state.productListing.language,
+    currency: state.productListing.currency,
     clientId: state.productListing.clientId,
     context: state.productListing.context,
     selectedFacets,